build-webos: add update capability to mcf
The update capability allow the various layers and other repos in the
build directory to be updated using the information in weboslayers.py.
Given this version of mcf is present, the user can download the latest
version of weboslayers.py, and run mcf with "-U" option. For example

./mcf -p 16 -b 64 -U qemuarm

Open-webOS-DCO-1.0-Signed-off-by: Iyad Qumei <iyad.qumei@palm.com>

Change-Id: Ib8d05f23bf1e7d2e07d9a21682700b28d75a4335